% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraph_constructor2.R
\name{extract_markers}
\alias{extract_markers}
\title{Extracts markers from cell population phenotypes}
\usage{
extract_markers(phen)
}
\arguments{
\item{phen}{A vector of cell population phenotypes.}
}
\value{
A vector of unique markers
}
\description{
Extracts all unique markers from cell population phenotypes
}
\seealso{
\code{\link[stringr]{str_split}}
}
