% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraphSubset_constructor.R
\name{fg_clean_phen}
\alias{fg_clean_phen}
\title{Reformats phenotype}
\usage{
fg_clean_phen(phen, markers = NULL)
}
\arguments{
\item{phen}{Vector of cell population phenotype names as character strings.}

\item{markers}{markers extracted from \code{phen}.}
}
\value{
Vector with the same length as \code{phen} containing reformatted
and not necessarily changed cell population phenotype names.
}
\description{
Reformats cell population phenotypes into flowGraph format
}
\examples{

# fg_clean_phen(c("A+_B+","B+_notC","A-_C"))

}
\seealso{
\code{\link[stringr]{str_extract}},\code{\link[stringr]{str_split}}
}
