% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_flowgraph_modifiers.R
\name{fg_clear_features}
\alias{fg_clear_features}
\title{Clears all featuresin a flowGraph object.}
\usage{
fg_clear_features(fg)
}
\arguments{
\item{fg}{flowGraph object.}
}
\value{
flowGraph object with only the \code{count} \code{node} feature.
}
\description{
Returns a flowGraph object with only the \code{count} feature.
}
\examples{

 no_cores <- 1
 data(fg_data_pos30)
 fg <- flowGraph(fg_data_pos30$count, class=fg_data_pos30$meta$class,
                 no_cores=no_cores)

 fg <- fg_clear_features(fg)
 fg_get_summary_desc(fg)

}
\seealso{
\code{\link[flowGraph]{flowGraph-class}}
}
