% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{fg_data_fca}
\alias{fg_data_fca}
\title{fg_data_fca}
\format{
A list containing the following elements
 derived from the flowCAP-II AML data set for cell populations up to layer 3.
\itemize{
  \item{\code{count}: A numeric sample x cell population node
  matrix with cell count values.}
  \item{\code{meta}: A data frame containing meta information on
   samples in \code{count}; it contains columns:}
  \itemize{
      \item{\code{class}: a string indicating whether a sample is
       from a "control" or "aml" subject.}
      \item{\code{id}: a string containing sample id's.}
      \item{\code{train}: a logical variable indicating whether
       a sample is from the train or test set.}
      \item{\code{subject}: a numeric variable containing the
       id of the subject from whom the sample came from.}
      \item{\code{tube}: the tube or panel number;
       all samples in this data set is
       analyzed under the 6th panel.}
  }
}
}
\source{
{
  \insertRef{aghaeepour2013critical}{flowGraph}
}
}
\usage{
fg_data_fca
}
\description{
fg_data_fca
}
\keyword{datasets}
