% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{fg_data_pos30}
\alias{fg_data_pos30}
\title{fg_data_pos30}
\format{
A list containing the following elements
 for a positive control data set with
 markers A, B, C, D; note it was made with two and three thresholds
 for markers A and B to test functions with multiple thresholds
 (this is a positive
 control data set where nodes A+..B+..C+ increased by 50%).
\itemize{
  \item{\code{count}: A numeric sample x cell population node
  matrix with cell count values}
  \item{\code{meta}: A data frame containing meta information on
  samples in \code{count}; it contains columns:}
  \itemize{
      \item{\code{id}: a string containing sample id's.}
      \item{\code{class}: a string indicating whether a sample is
       from a "control" or "exp" (experiment) subject.}
  }
}
}
\usage{
fg_data_pos30
}
\description{
fg_data_pos30
}
\keyword{datasets}
