% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_flowgraph_accessors.R
\name{fg_get_meta}
\alias{fg_get_meta}
\title{Retrieves sample meta.}
\usage{
fg_get_meta(fg)
}
\arguments{
\item{fg}{flowGraph object.}
}
\value{
A data frame containing sample meta data.
}
\description{
Retrieves sample meta from a given flowGraph object.
}
\examples{

 no_cores <- 1
 data(fg_data_pos30)
 fg <- flowGraph(fg_data_pos30$count, class=fg_data_pos30$meta$class,
                 prop=FALSE, specenr=FALSE,
                 no_cores=no_cores)
 head(fg_get_meta(fg))

}
\seealso{
\code{\link[flowGraph]{flowGraph-class}}
 \code{\link[flowGraph]{fg_replace_meta}}
}
