% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowgraph_class.R
\name{show}
\alias{show}
\alias{show,flowGraph-method}
\title{flowGraph object accessor}
\usage{
\\method{show}{flowGraph}(x)
}
\arguments{
\item{object}{Object of class \code{flowGraph}.}
}
\value{
A list containing the two data frames from the \code{feat_desc} and
 \code{summary_desc} slots of a \code{flowGraph} object.
}
\description{
Shows a summary of a given flowGraph object.
}
\details{
Accessor method for the data frames containing information on features
and feature summaries in the \code{feat_desc} and \code{summary_desc}
slots of a \code{flowGraph} object.
}
\examples{

 no_cores <- 1
 data(fg_data_pos30)
 fg <- flowGraph(fg_data_pos30$count, meta=fg_data_pos30$meta,
                 prop=FALSE, specenr=FALSE,
                 no_cores=no_cores)
 show(fg)

}
\seealso{
\code{\link[flowGraph]{flowGraph-class}}
}
\keyword{methods}
