% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraph_constructor2.R
\name{flowGraph2_summary_adjust}
\alias{flowGraph2_summary_adjust}
\title{Default for flowGraph2's summary_adjust}
\usage{
flowGraph2_summary_adjust()
}
\value{
Default list parameter flowGraph2's \code{summary_adjust} parameter.
}
\description{
Default input for flowGraph2's \code{summary_adjust} parameter.
}
\examples{

 flowGraph2_summary_adjust()

}
