% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraph_constructor2.R
\name{flowGraph2_summary_pars}
\alias{flowGraph2_summary_pars}
\title{Default for flowGraph2's summary_pars}
\usage{
flowGraph2_summary_pars()
}
\value{
Default list parameter flowGraph2's \code{summary_pars} parameter.
}
\description{
Default input for flowGraph2's \code{summary_pars} parameter.
}
\examples{

 flowGraph2_summary_pars()

}
