% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraphSubset_constructor.R
\name{flowGraphSubset_summary_pars}
\alias{flowGraphSubset_summary_pars}
\title{Default for flowGraphSubset's summary_pars}
\usage{
flowGraphSubset_summary_pars()
}
\value{
Default list parameter flowGraphSubset's \code{summary_pars} parameter.
}
\description{
Default input for flowGraphSubset's \code{summary_pars} parameter.
}
\examples{

 flowGraphSubset_summary_pars()

}
