% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraphSubset_constructor.R
\name{fpurrr_map}
\alias{fpurrr_map}
\title{Wrapper for map}
\usage{
fpurrr_map(x, f, no_cores = 1, prll = TRUE, ...)
}
\arguments{
\item{x}{Variable to recurse over; must be indices!}

\item{f}{Function to recurse over.}

\item{no_cores}{Number of cores to use; \code{future} must have
been ran already.}

\item{prll}{If set to FALSE, forces use of purrr::map instead of
furrr::future_map, Default: TRUE}

\item{...}{Other parameters used by \code{f}.}
}
\value{
Unnested named list.
}
\description{
Wrapper for \code{purrr::map} and \code{furrr::future_map}
 to handle parallel-ization
}
\details{
Wrapper for \code{purrr::map} and \code{furrr::future_map} to handle
parallel-ization easily; note that \code{future} must have been ran already
outside of the function and outputs will always be a list.
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[furrr]{future_map}}
}
