% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraph_constructor2.R
\name{get_eprop}
\alias{get_eprop}
\title{Gets edge proportions of a given edge matrix}
\usage{
get_eprop(edf_, ep, mp_, no_cores = 1)
}
\arguments{
\item{edf_}{edge x from&to data frame containing edges and their from and to
cell population phenotypes.}

\item{ep}{sample x edge (parent_child) matrix with edge proportions.}

\item{mp_}{sample x phenotype matrix with proportions.}

\item{no_cores}{Number of cores to use, Default: 1}
}
\value{
\code{ep} with only the specific columns (edges) requested; also updates
\code{ep} globally.
}
\description{
Gets the edge proportions of the edges in edge matrix
 \code{edf_} and updates \code{ep} edge proportion matrix if it didn't
 contain the requested information.
}
