% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraph_constructor2.R
\name{ms_create}
\alias{ms_create}
\title{Calcuate SpecEnr from proportion and expected proportion}
\usage{
ms_create(mp_, me_)
}
\arguments{
\item{mp_}{Numerical sample x cell population matrix w/ proportions.}

\item{me_}{Numerical sample x cell population matrix w/ expected proportions.}
}
\value{
Numerical sample x cell population matrix w/ SpecEnr.
}
\description{
FUNCTION_DESCRIPTION
}
