% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraph_constructor2.R
\name{ms_psig}
\alias{ms_psig}
\title{Determines which phenotypes are statistically significant}
\usage{
ms_psig(
  ms_,
  summary_pars,
  summary_adjust,
  test_cust,
  test_custom,
  lyrno,
  mp_,
  me_
)
}
\arguments{
\item{ms_}{sample x phenotype SpecEnr matrix}

\item{summary_pars}{See \code{flowGraph2}.}

\item{summary_adjust}{See \code{flowGraph2}.}

\item{test_cust}{Final significance test function.}

\item{test_custom}{Raw significance test function.}

\item{mp_}{sample x phenotype proportion matrix.}

\item{me_}{sample x phenotype expected proportion matrix.}

\item{lyrs}{Total number of layers in hierarchy, used for byLayer p-value
adjustment.}
}
\value{
A logical vector the same length as the number of columns in
 \code{ms_} indicating whether or not each phenotype is significant;
 used only for the fast version of flowGraph to determine whether or not
 to keep testing the phenotypes' children.
}
\description{
Determines which phenotypes are statistically significant based
 on SpecEnr.
}
