% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helpers.R
\name{ncores}
\alias{ncores}
\title{Prepares number of cores for parallel backend.}
\usage{
ncores(no_cores = 1)
}
\arguments{
\item{no_cores}{An integer indicating how many cores to parallelize on.}
}
\value{
An integer indicating how many cores to parallelize on.
}
\description{
Prepares number of cores needed for parallel backend.
}
\details{
Given the number of cores a user wishes to parallelize processes on
 \code{no_cores}, \code{ncores} ensures this value does not
 exceed the actual number of cores the user's computer contains.
}
\examples{

 # NOT EXPORTED
 no_cores <- 100
 flowGraph:::ncores(no_cores)

}
\seealso{
\code{\link[parallel]{detectCores}}
}
