% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helpers.R
\name{summary_table}
\alias{summary_table}
\title{Summarizes a numeric matrix.}
\usage{
summary_table(m, feat_type = "")
}
\arguments{
\item{m}{A numeric matrix.}

\item{feat_type}{Name of the matrix \code{m}.}
}
\value{
A data frame containing one row summarizing \code{m};
 see \code{\link[flowGraph]{fg_get_feature_desc}}.
}
\description{
Summarizes a numeric matrix.
}
\examples{

 summary_table(matrix(rnorm(12),nrow=3), feat_type='random')

}
