% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_flowGraph_constructor2.R
\name{test_c}
\alias{test_c}
\title{Converts input into a significance test function}
\usage{
test_c(test_custom)
}
\arguments{
\item{test_custom}{a string \code{c("t", "wilcox","ks","var","chisq")} or a
function.}
}
\value{
a statistical significance test function.
}
\description{
Converts input into a significance test function
}
\seealso{
\code{\link[stats]{t.test}},\code{\link[stats]{wilcox.test}},
 \code{\link[stats]{ks.test}},\code{\link[stats]{var.test}},
 \code{\link[stats]{chisq.test}}
}
