% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helpers.R
\name{time_output}
\alias{time_output}
\title{Outputs elapsed time.}
\usage{
time_output(start, msg = "")
}
\arguments{
\item{start}{A time variable of class \code{POSIXct}, \code{POSIXt}.}

\item{msg}{A string with a message to print out after the elapsed time.}
}
\value{
Prints to console, the time from which process
 started \code{start} - ended, and > time elapsed from
 \code{start} until now.
}
\description{
Given a time, prints the time elapsed from that time until now.
}
\examples{

 start <- Sys.time()
 flowGraph:::time_output(start,'start - now > time elapsed')

}
