% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_flowgraph_features.R
\name{tmm}
\alias{tmm}
\title{Normalizes inter-row variation.}
\usage{
tmm(
  x,
  x0 = NULL,
  lib.size,
  refColumn,
  cutoff = Inf,
  pngnames = NULL,
  mains = NULL,
  no_cores = 1,
  samplesOnCol = FALSE,
  Acutoff = -10,
  doWeighting = FALSE,
  logratioTrim = 0.3,
  sumTrim = 0.05,
  minlogR = 1e-06
)
}
\arguments{
\item{x}{A numeric matrix.}

\item{x0}{A numeric matrix, usually the same as \code{x}, no need to specify.
This is only for plotting purposes, so if the user wants more or less
points on the plots than those in \code{x} or those used to conduct
normalization.}

\item{lib.size}{A numeric matrix the same length as the number
of samples is \code{x}.}

\item{refColumn}{An integer indicating which sample in \code{x} should act as
the reference sample in normalization.}

\item{cutoff}{A double indicating a threshold. Given that \code{tmm} finds
both the trimmed mean and the density peak of inter-sample count ratios,
it by default sets the normalization factor as the trimmed mean ---
unless the difference between trimmed mean
and the peak is more than \code{cutoff},
then \code{tmm} will use the peak value as the normalization factor instead.}

\item{pngnames}{A string vector containing full PNG file paths to where
\code{tmm} can save plots; set as \code{NULL} to not save plots.}

\item{mains}{A string vector cntaining plot titles for plots corresponding to
those saved in \code{pngnames}; set as \code{NULL} if not saving plots.}

\item{no_cores}{An integer indicating how many cores to parallelize on.}

\item{samplesOnCol}{A logical variable indicating whether samples or the
dimension user wants to normalize on is on the columns and
not on the rows in \code{x}.}

\item{Acutoff}{An integer or the cutoff on "A" values to use before trimming.}

\item{doWeighting}{A logical variable indicating whether to compute
(asymptotic binomial precision) weights.}

\item{logratioTrim}{A double or the amount of trim to use on log-ratios
("M" values).}

\item{sumTrim}{A double or the amount of trim to use on the
combined absolute levels ("A" values).}

\item{minlogR}{A double or the minimum value of the logged ratio
used to calculate the normalization factor.}
}
\value{
A list containing:
\itemize{
  \item{\code{f}: a numeric vector with the normalization factor for
   each sample in matrix \code{x}.}
  \item{\code{fdiff}: a numeric vector indicating the difference between
    the normalization factors for each sample in matrix \code{x}
    and the ratio peak value.}
}
}
\description{
Normalizes inter-row variation given a numeric matrix.
 Note this function uses a modified version of
 the \code{calcNormFactors} function from the \code{edgeR} package.
 This function is only used in \code{\link[flowGraph]{fg_feat_node_norm}}
 and is not recommended for use on its own.
}
\details{
This function calculates factors for trimmed mean normalization
 of count data.
}
\examples{

 # NOT EXPORTED
 no_cores <- 1
 data(fg_data_pos30)
 tmm_factors <- flowGraph:::tmm(x=fg_data_pos30$count,
                                lib.size=fg_data_pos30$count[,1],
                                refColumn=1, no_cores=no_cores)

}
\references{
\insertRef{robinson2010scaling}{flowGraph}
}
\seealso{
\code{\link[flowGraph]{fg_feat_node_norm}}
}
