% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helpers.R
\name{tstr}
\alias{tstr}
\title{Formats time into string.}
\usage{
tstr(time)
}
\arguments{
\item{time}{A time variable of class \code{POSIXct}, \code{POSIXt}.}
}
\value{
Time formatted as a string; used in \code{time_output} function.
}
\description{
Formats time into a string HH:MM:SS given time zone.
}
\examples{
 # NOT EXPORTED
 flowGraph:::tstr(Sys.time())

}
